Move management to its own API endpoint
