+++ maediprojects/views/__init__.py
@@ -1,1 +1,1 @@
-from maediprojects.views import activities, api, iati, codelists, users, exports, reports
+from maediprojects.views import activities, api, iati, codelists, users, exports, reports, management
